Fix BubblePark.gameWorldWidth return type

+++ composeApp/src/commonMain/kotlin/fr/iutlens/mmi/demo/BubblePark.kt
@@ -126,7 +126,7 @@ class BubblePark : GameData() {
     private lateinit var distanceMap: DistanceMap
 
     val gameWorldWidth: Float
-        get() = tileArea.tileMap.geometry.sizeX * tileArea.w
+        get() = (tileArea.tileMap.geometry.sizeX * tileArea.w).toFloat()
 
     var onLevelEnd: ((hasNextLevel: Boolean) -> Unit)? = null
     var levelIndex by mutableStateOf(0)
